+++ src/engine/components/numberGenerator/index.js
@@ -5,10 +5,12 @@
  * @flow
  */
 import { Text, TextStyle } from 'pixi.js';
-import { TweenLite, Linear } from 'gsap';
+import { TweenLite, TimelineLite, Linear } from 'gsap';
 
-import { HALF, ACTOR_MOVEMENT_DURATION } from 'constants/numbers';
+import { HALF, ONE, ZERO, TWO, ANCHOR_CENTER, ACTOR_MOVEMENT_DURATION } from 'constants/numbers';
 import { UnableToLeaveTheNumericLine } from 'engine/helpers/errors';
+import { EASE_BE_HAPPY } from 'engine/helpers/customEases';
+import { getRandomFloat } from 'engine/helpers/randomConfigurations';
 
 const EIGHT = 8;
 
@@ -30,6 +32,7 @@ export type NumberActor = {|
   view: Text,
   position: string,
   finalPosition: string,
+  beHappy(state: BeHappyState): void,
   changeActor(number: number): void,
   hasEnteredToNumericLine(): Promise<void>,
   resetPosition(): void,
@@ -37,7 +40,35 @@ export type NumberActor = {|
 |};
 export type StaticNumberActor = {|
   view: Text,
+  beHappy(state: BeHappyState): void,
 |};
+export type BeHappyState = 'start' | 'stop';
+
+const beHappyConfig = (
+  view: Text,
+) => {
+  const timeline = new TimelineLite({
+    onComplete: () => {
+      timeline.restart();
+    },
+    paused: true,
+  });
+  const delay = getRandomFloat(ZERO, TWO);
+
+  timeline.to(view, ONE, {
+    y: 0,
+    ease: EASE_BE_HAPPY,
+    delay,
+  });
+
+  return (state: BeHappyState): void => {
+    if (state === 'start') {
+      timeline.restart();
+    } else {
+      timeline.stop();
+    }
+  };
+};
 
 export const staticNumberGenerator = (number: number, size: number): StaticNumberActor => {
   const style = new TextStyle({
@@ -46,11 +77,12 @@ export const staticNumberGenerator = (number: number, size: number): StaticNumbe
   });
   const view = new Text(number.toString(), style);
 
-  view.anchor.x = view.anchor.y = 0.5;
+  view.anchor.set(ANCHOR_CENTER);
   view.x = view.y = size / HALF;
 
   return {
     view,
+    beHappy: beHappyConfig(view),
   };
 };
 
@@ -147,7 +179,7 @@ const numberGenerator = (
   const view = new Text(number.toString(), style);
   const initialPosition = position.split(',').map((string: string): number => (parseInt(string)));
 
-  view.anchor.x = view.anchor.y = 0.5;
+  view.anchor.set(ANCHOR_CENTER);
 
   view.x = initialPosition[0] * (size + margin) + size / HALF + margin;
   view.y = initialPosition[1] * (size + margin) + size / HALF + margin;
@@ -156,6 +188,7 @@ const numberGenerator = (
     view,
     position,
     finalPosition,
+    beHappy: beHappyConfig(view),
     changeActor: changeActorConfig(view),
     hasEnteredToNumericLine: hasEnteredToNumericLineConfig(view, size, margin),
     resetPosition: resetPositionConfig(view, initialPosition, size, margin),